refactor(provider-settings): share configuration and model workflows
