+++ src/frontend/features/settings/provider/detail/ProviderDetailScreen.tsx
@@ -1,56 +1,21 @@
 import PlusIcon from '@cherrystudio/app-icons/icons/plus';
 import RefreshCwIcon from '@cherrystudio/app-icons/icons/refresh-cw';
-import {
-  Alert,
-  Button,
-  ContentState,
-  Spinner,
-  useAlert,
-  useToast,
-} from '@cherrystudio/ui/components';
+import { Alert, Button, ContentState, Spinner, useToast } from '@cherrystudio/ui/components';
 import { Redirect, useLocalSearchParams, useRouter } from 'expo-router';
 import { useCallback, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Keyboard, StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Text, View } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-controller';
 
 import { RouteHeader, type HeaderToolbarAction } from '@/frontend/appShell/header';
 import { ProviderBrandAvatar } from '@/frontend/components/Avatar';
 import { InlineSearch, useInlineSearch } from '@/frontend/components/InlineSearch';
 import { SelectionToolbar } from '@/frontend/components/Selection';
-import { useQuery } from '@/frontend/data';
 import { keyboardBottomOffset } from '@/frontend/utils/constants';
-import type { UpdateProviderInput } from '@/shared/data/api/schemas/providers';
 import type { Model } from '@/shared/data/types/model';
 
-import {
-  buildApiKeyEntriesFromInput,
-  buildApiKeysInputFromEntries,
-  buildProviderPrimaryBaseUrlUpdates,
-  buildProviderTextEndpointUpdates,
-  getEffectiveAuthConfig,
-  isFullyCustomProvider,
-  normalizeApiKeyEntries,
-  ProviderApiServiceSaveError,
-  shouldShowApiKeys,
-  useProviderApiServiceQueries,
-  useProviderApiServiceSheetClose,
-} from '../apiService';
-import {
-  CUSTOM_PROVIDER_TEXT_ENDPOINT_TYPES,
-  findInvalidCustomProviderEndpointUrl,
-  hasConfiguredCustomProviderTextEndpoint,
-} from '../apiService/utils/providerApiServiceEndpointRules';
-import {
-  createEmptyProviderFormValues,
-  createProviderFormValues,
-  providerDefaultEndpointNeedsRepair,
-  ProviderForm,
-  providerFormAvatarSize,
-  resolveProviderFormEndpointTypes,
-  useProviderFormDraft,
-} from '../components/ProviderForm';
-import { useProviderAvatar, useProviderAvatarActions } from '../hooks/useProviderAvatar';
+import { useProviderApiServiceSheetClose, useProviderConfigurationForm } from '../apiService';
+import { ProviderForm, providerFormAvatarSize } from '../components/ProviderForm';
 import { useProviderDeletion } from '../hooks/useProviderDeletion';
 import { useProviderSetup } from '../hooks/useProviderSetup';
 import { ProviderModelCheckSection } from '../models/components/ProviderModelCheckSection';
@@ -92,21 +57,27 @@ function ProviderDetailSettings({
 }) {
   const { t } = useTranslation();
   const router = useRouter();
-  const { alert } = useAlert();
   const { toast } = useToast();
-  const providerAvatars = useProviderAvatarActions();
   const { tab } = useLocalSearchParams<{ tab?: string }>();
   const activeTab: ProviderDetailTab = tab === 'models' ? 'models' : 'configuration';
   const { isPreparing, openSetup } = useProviderSetup();
   const [isSyncPromptOpen, setIsSyncPromptOpen] = useState(false);
   const [modelPurpose, setModelPurpose] = useState<ProviderModelPurpose>('all');
-  const [isSaving, setIsSaving] = useState(false);
-  const { models, modelsQuery, provider, providerQuery } = useProviderDetailSettings(providerId);
-  const isCustomProvider = isFullyCustomProvider(provider);
-  const allProviderModelsQuery = useQuery('/models', {
-    enabled: Boolean(providerId),
-    query: { providerId },
-  });
+  const { models, modelsQuery } = useProviderDetailSettings(providerId);
+  const {
+    apiKeys,
+    canSubmit: canSubmitProvider,
+    createInitialValues: createInitialFormValues,
+    form,
+    isCustomProvider,
+    isLoading: isProviderDetailLoading,
+    isSaving,
+    modelsQuery: allProviderModelsQuery,
+    provider,
+    providerQuery,
+    requestSave,
+    showApiKey: showApiKeys,
+  } = useProviderConfigurationForm(providerId);
   const allProviderModels = useMemo(
     () => allProviderModelsQuery.data ?? [],
     [allProviderModelsQuery.data],
@@ -137,66 +108,7 @@ function ProviderDetailSettings({
   const management = useProviderModelManagement(providerId, managedModels, listedModels);
   const isModelListFiltered = isModelSearchActive || effectiveModelPurpose !== 'all';
   const showsModelPurposeTabs = hasMultipleProviderModelPurposes(modelPurposeCounts);
-  const {
-    apiKeys,
-    apiKeysQuery,
-    authConfig,
-    authConfigQuery,
-    replaceApiKeysMutation,
-    saveProviderMutation,
-  } = useProviderApiServiceQueries(providerId);
-  const storedAvatarUri = useProviderAvatar(providerId);
-  const defaultEndpointNeedsRepair = provider
-    ? providerDefaultEndpointNeedsRepair(provider)
-    : false;
-  const endpointTypes = useMemo(
-    () => (provider ? resolveProviderFormEndpointTypes(provider) : []),
-    [provider],
-  );
-  const showApiKeys = shouldShowApiKeys(
-    getEffectiveAuthConfig(authConfig, provider).type,
-    provider,
-  );
-  const apiKeysInput = useMemo(
-    () => buildApiKeysInputFromEntries(normalizeApiKeyEntries(apiKeys ?? [])),
-    [apiKeys],
-  );
-  // Gate on every required query so the content reaches its final structure on the first frame.
-  // Inserting the Base URL / API keys blocks a commit later shifts the model toolbar
-  // under a finger that already aimed at it.
-  const isProviderDetailLoading =
-    providerQuery.isPending ||
-    apiKeysQuery.isPending ||
-    authConfigQuery.isPending ||
-    allProviderModelsQuery.isPending;
-  const createInitialFormValues = useCallback(
-    () =>
-      provider
-        ? createProviderFormValues({
-            apiKey: apiKeysInput,
-            avatarUri: storedAvatarUri ?? null,
-            provider,
-          })
-        : createEmptyProviderFormValues(),
-    [apiKeysInput, provider, storedAvatarUri],
-  );
-  const form = useProviderFormDraft({
-    createInitialValues: createInitialFormValues,
-    defaultEndpointNeedsRepair,
-    endpointTypes,
-    initiallyDirty: defaultEndpointNeedsRepair,
-    isSubmitting: isSaving,
-    normalizeCustomEndpoints: isCustomProvider,
-    sourceKey: !isProviderDetailLoading && provider ? provider.id : '',
-  });
   const { meta: formMeta, state: formState } = form;
-  const customEndpointError = isCustomProvider
-    ? findInvalidCustomProviderEndpointUrl(formState.endpointUrls)
-    : null;
-  const canSubmitProvider =
-    formMeta.canSubmit &&
-    (!isCustomProvider ||
-      (hasConfiguredCustomProviderTextEndpoint(formState.endpointUrls) && !customEndpointError));
   const { allowNavigation, requestClose } = useProviderApiServiceSheetClose({
     hasUnsavedChanges: formMeta.isDirty && !management.isSelecting,
     isSaving: isSaving || management.isDeleting,
@@ -209,171 +121,13 @@ function ProviderDetailSettings({
   }, [provider, requestDelete]);
   const handleSave = useCallback(
     (onSaved?: () => void) => {
-      if (!provider || !providerId || !canSubmitProvider || !formMeta.isDirty) {
-        return;
-      }
-
-      const trimmedName = formState.name.trim();
-      let updates: UpdateProviderInput = {
-        name: trimmedName,
-      };
-      const baseUrlEndpoint = endpointTypes[0];
-      let savedEndpointUrls = formState.endpointUrls;
-
-      if (isCustomProvider) {
-        try {
-          const endpointUpdates = buildProviderTextEndpointUpdates({
-            defaultChatEndpoint: formState.defaultChatEndpoint,
-            endpointUrls: formState.endpointUrls,
-            provider,
-          });
-          updates = { ...updates, ...endpointUpdates };
-          savedEndpointUrls = Object.fromEntries(
-            CUSTOM_PROVIDER_TEXT_ENDPOINT_TYPES.map((endpointType) => [
-              endpointType,
-              formState.endpointUrls[endpointType]?.trim() ?? '',
-            ]),
-          );
-        } catch (error) {
-          alert.show(
-            error instanceof ProviderApiServiceSaveError && error.code === 'missing-text-endpoint'
-              ? {
-                  description: t('settings.provider.apiService.textEndpointRequired'),
-                  title: t('settings.provider.apiService.textEndpointsTitle'),
-                }
-              : {
-                  description: t('settings.provider.apiService.invalidBaseUrlMessage'),
-                  title: t('settings.provider.apiService.invalidBaseUrlTitle'),
-                },
-          );
-          return;
-        }
-      } else if (baseUrlEndpoint) {
-        try {
-          updates = {
-            ...updates,
-            ...buildProviderPrimaryBaseUrlUpdates({
-              baseUrl: formState.endpointUrls[baseUrlEndpoint] ?? '',
-              provider,
-            }),
-          };
-        } catch (error) {
-          alert.show(
-            error instanceof ProviderApiServiceSaveError
-              ? {
-                  description: t('settings.provider.apiService.invalidBaseUrlMessage'),
-                  title: t('settings.provider.apiService.invalidBaseUrlTitle'),
-                }
-              : { title: t('settings.provider.apiService.saveFailed') },
-          );
-          return;
-        }
-        savedEndpointUrls = {
-          ...formState.endpointUrls,
-          [baseUrlEndpoint]: (formState.endpointUrls[baseUrlEndpoint] ?? '').trim(),
-        };
-      }
-
-      const removedEndpointTypes = isCustomProvider
-        ? CUSTOM_PROVIDER_TEXT_ENDPOINT_TYPES.filter(
-            (endpointType) =>
-              provider.endpointConfigs?.[endpointType]?.baseUrl?.trim() &&
-              !updates.endpointConfigs?.[endpointType]?.baseUrl?.trim(),
-          )
-        : [];
-      const referencedModelCount = allProviderModels.filter((model) => {
-        const endpointType = model.endpointTypes?.[0];
-        return endpointType
-          ? removedEndpointTypes.some((removed) => removed === endpointType)
-          : false;
-      }).length;
-      if (referencedModelCount > 0) {
-        alert.show({
-          description: t('settings.provider.apiService.endpointInUseMessage', {
-            count: referencedModelCount,
-          }),
-          title: t('settings.provider.apiService.endpointInUseTitle'),
-        });
-        return;
-      }
-
-      const nextApiKeys = buildApiKeyEntriesFromInput(formState.apiKey, apiKeys ?? []);
-      const shouldSaveApiKeys = showApiKeys && formState.apiKey !== apiKeysInput;
-      const persistUpdates = () => {
-        Keyboard.dismiss();
-        setIsSaving(true);
-        void Promise.all([
-          saveProviderMutation.mutateAsync(updates),
-          shouldSaveApiKeys ? replaceApiKeysMutation.mutateAsync(nextApiKeys) : Promise.resolve(),
-        ])
-          .then(async () => {
-            if (formState.avatarUri !== (storedAvatarUri ?? null)) {
-              if (formState.avatarUri) {
-                await providerAvatars.persist(providerId, formState.avatarUri);
-              } else {
-                providerAvatars.remove(providerId);
-              }
-            }
-
-            form.actions.reset({
-              ...formState,
-              apiKey: shouldSaveApiKeys
-                ? buildApiKeysInputFromEntries(nextApiKeys)
-                : formState.apiKey,
-              defaultChatEndpoint: updates.defaultChatEndpoint ?? formState.defaultChatEndpoint,
-              endpointUrls: savedEndpointUrls,
-              name: trimmedName,
-            });
-            toast.show({ label: t('settings.provider.toast.saved'), variant: 'success' });
-            onSaved?.();
-          })
-          .catch(() => {
-            toast.show({ label: t('settings.provider.apiService.saveFailed'), variant: 'danger' });
-          })
-          .finally(() => setIsSaving(false));
-      };
-      const followingModelCount = allProviderModels.filter(
-        (model) => !model.endpointTypes?.[0],
-      ).length;
-      if (
-        isCustomProvider &&
-        provider.defaultChatEndpoint !== updates.defaultChatEndpoint &&
-        followingModelCount > 0
-      ) {
-        alert.confirm({
-          confirmLabel: t('common.save'),
-          description: t('settings.provider.apiService.defaultEndpointChangeMessage', {
-            count: followingModelCount,
-          }),
-          onConfirm: persistUpdates,
-          title: t('settings.provider.apiService.defaultEndpointChangeTitle'),
-        });
-        return;
-      }
-
-      persistUpdates();
+      if (!formMeta.isDirty) return;
+      requestSave(() => {
+        toast.show({ label: t('settings.provider.toast.saved'), variant: 'success' });
+        onSaved?.();
+      });
     },
-    [
-      alert,
-      allProviderModels,
-      apiKeys,
-      apiKeysInput,
-      canSubmitProvider,
-      endpointTypes,
-      form.actions,
-      formMeta.isDirty,
-      formState,
-      isCustomProvider,
-      provider,
-      providerAvatars,
-      providerId,
-      replaceApiKeysMutation,
-      saveProviderMutation,
-      showApiKeys,
-      storedAvatarUri,
-      t,
-      toast,
-    ],
+    [formMeta.isDirty, requestSave, t, toast],
   );
   const configurationSaveActions = useMemo<HeaderToolbarAction[]>(
     () => [
