+++ src/frontend/features/settings/provider/detail/modelAdd/types.ts
@@ -1,0 +1,7 @@
+import type { Provider } from '@/shared/data/types/provider';
+
+export type ProviderModelTaskProps = {
+  provider: Provider;
+  returnTo?: string;
+  shouldEnableProvider: boolean;
+};
