+++ src/frontend/features/settings/provider/models/README.md
@@ -4,11 +4,11 @@ This module owns provider model listing, connectivity checks, synchronization, a
 
 ## Public Interface
 
-- Model list leaf components and `useProviderModelGroups` are exported from `index.ts`.
+- Provider pages consume the model components, workflow hooks, and pure helpers owned here.
 
 ## Organization
 
-- `components/` contains provider model list UI pieces.
+- `components/` contains model row content and token-limit inputs shared by management and model tasks.
 - `hooks/` owns displayed group state plus add/sync workflows.
 - `utils/` contains pure grouping and filtering helpers, synchronization previews, and the check's
   selection resolvers.
@@ -27,6 +27,8 @@ no changes selected. Errors are classified into configuration, authentication, n
 unavailable-directory, and rate-limit outcomes. The screen owns inline feedback and routes to
 configuration repair or an independent manual-add task with the same setup intent and return target.
 An empty directory also offers manual creation. Removal results report protected skipped models.
+The manual form and synchronization task mount independently under `detail/modelAdd/components/`;
+the synchronization preview lives with that task, while the legacy pull page only redirects.
 
 The management list uses CherryUI context menus and a scroll boundary. A tap opens details; a long
 press offers details, editing, selection, and deletion. Selection disables navigation and endpoint
@@ -37,3 +39,7 @@ Management deletion uses the transactional model DELETE endpoint, not synchroniz
 which intentionally protects custom models. Default model assignments block deletion and link to
 model settings. Agent bindings are cleared by the existing database foreign key; model and Agent
 queries are refreshed, including inactive Agent detail caches. The single request limit is 1000 IDs.
+
+Model rows share one content layout across browsing and selection, including visible and accessible
+availability labels. Endpoint options are shared between the list picker and editor; the list saves
+immediately, while the editor keeps the choice in its draft until Save.
